% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_markow.R
\name{as_kmer_prob}
\alias{as_kmer_prob}
\title{Convert a data frame into a k-mer probability table}
\usage{
as_kmer_prob(x)
}
\arguments{
\item{x}{The input data frame.}
}
\value{
A k-mer table.
}
\description{
This function takes a data frame and converts it into a k-mer table. The input data frame must have columns named "kmer" and "count".
}
\examples{
data <- data.frame(kmer = c("A", "T", "C"), count = c(2, 5, 1))
as_kmer_table(data)

}
