% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_markov.R
\name{build_markov}
\alias{build_markov}
\title{A function to return conditional probabilities from a Markov model of a given order.}
\usage{
build_markov(fasta_file, order_markov)
}
\arguments{
\item{fasta_file}{A fasta file, containing a list of genetic sequences.}

\item{order_markov}{Order of the Markov model.}
}
\value{
A data frame with the following columns:
  \describe{
    \item{kmer}{The kmer sequence.}
    \item{knowing}{The prefix of the kmer (substring from the beginning to k-1).}
    \item{emission}{The last character of the kmer (substring from k to the end).}
    \item{count}{The count of occurrences of the kmer.}
    \item{knowing_prob}{The probability of the prefix (knowing) occurring.}
    \item{knowing_lprob}{The logarithm of the knowing_prob.}
    \item{prob}{The conditional probability of the kmer given the prefix.}
    \item{lprob}{The logarithm of the prob.}
  }
}
\description{
A function to return conditional probabilities from a Markov model of a given order.
}
\examples{
To build a Markov model of order 3, on a fasta file named "mus_cpg_app.fa" containing sequences.
proba_markov <- build_markov("mus_cpg_app.fa", 3)

}
