% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_table.R
\name{every_kmer_table}
\alias{every_kmer_table}
\title{Function to create a table with all possible k-mers and their counts}
\usage{
every_kmer_table(k, pseudocount = 1)
}
\arguments{
\item{k}{The length of the k-mers}

\item{pseudocount}{The initial count for each k-mer (default = 1)}
}
\value{
A k-mer table with k-mers and their counts
}
\description{
Function to create a table with all possible k-mers and their counts
}
\examples{
every_kmer_table(3)
}
