% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{HMM_as_sequence}
\alias{HMM_as_sequence}
\title{Extracts every segment of a sequence as an individual sequence.}
\usage{
HMM_as_sequence(segmentation)
}
\arguments{
\item{segmentation}{: a tibble returned by the \code{viterbi} function}
}
\value{
a sequence tibble containing four columns
\itemize{
\item id : the identifier of the sequence
\item model : the model corresponding to that segment
\item from : the start position of the fragment in the segmented sequence
\item to : the end position of the fragment in the segmented sequence
\item sequence : the nucleotide sequence of the fragment
}
}
\description{
Extracts every segment of a sequence as an individual sequence.
}
