% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_markow.R
\name{is_kmer_proba}
\alias{is_kmer_proba}
\title{Check if an object is a kmer probability table}
\usage{
is_kmer_proba(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical value indicating if the object is a kmer probability table
}
\description{
Check if an object is a kmer probability table
}
