% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_table.R
\name{is_kmer_table}
\alias{is_kmer_table}
\title{Check if an object is a kmer table}
\usage{
is_kmer_table(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical value indicating if the object is a kmer table
}
\description{
Check if an object is a kmer table
}
