% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_markow.R
\name{kmer_proba}
\alias{kmer_proba}
\title{Function to calculate the conditional probability of each emitted nucleotide}
\usage{
kmer_proba(kmers, pseudocount = 1)
}
\arguments{
\item{kmers}{A vector of kmers.}

\item{pseudocount}{The pseudocount value to be used (default = 1).}
}
\value{
A data frame with the following columns:
  \describe{
    \item{kmer}{The kmer sequence.}
    \item{knowing}{The prefix of the kmer (substring from the beginning to k-1).}
    \item{emission}{The last character of the kmer (substring from k to the end).}
    \item{count}{The count of occurrences of the kmer.}
    \item{knowing_prob}{The probability of the prefix (knowing) occurring.}
    \item{knowing_lprob}{The logarithm of the knowing_prob.}
    \item{prob}{The conditional probability of the kmer given the prefix.}
    \item{lprob}{The logarithm of the prob.}
  }
}
\description{
This function calculates the probability of each kmer sequence based on its prefix and emission.
It takes a vector of kmers and an optional pseudocount value as input.
}
\examples{
kmers <- as_kmer_table(data.frame(kmer = c("AAA", "AAC", "AAG"), count = c(10, 20, 30)))
kmer_proba(kmers)

}
