% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_size.R
\name{kmer_size}
\alias{kmer_size}
\title{Get the kmer size of a kmer table}
\usage{
kmer_size(x)
}
\arguments{
\item{x}{A kmer table object}
}
\value{
The kmer size of the kmer table
}
\description{
Get the kmer size of a kmer table
}
