% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_table.R
\name{kmer_table}
\alias{kmer_table}
\title{Generate k-mer table from a given sequence}
\usage{
kmer_table(sequence, k)
}
\arguments{
\item{sequence}{The DNA sequence}

\item{k}{The length of the k-mers}
}
\value{
A k-mer table with k-mer counts
}
\description{
This function takes a DNA sequence and generates a k-mer table,
which counts the occurrence of each k-mer in the sequence.
}
\examples{
sequence <- "ATCGATCGATCG"
kmer_table(sequence, 3)

}
