% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_table.R
\name{kmer_table_seq}
\alias{kmer_table_seq}
\title{Generate k-mer table for a list of sequences}
\usage{
kmer_table_seq(sequences, k)
}
\arguments{
\item{sequences}{A data frame with a column named 'sequence' containing the sequences}

\item{k}{The value of k for generating k-mer table}
}
\value{
A table of k-mer frequencies for each sequence combined into a single table
}
\description{
This function takes a list of sequences and a value of k as input and returns a table of k-mer frequencies for each sequence combined into a single table.
The function expects a data frame with a column named 'sequence' that contains the sequences.
}
\examples{
sequences <- data.frame(sequence = c("ACGT", "CGTA"))
kmer_table_seq(sequences, 2)

}
