% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_likelihood.R
\name{markov_likelihood}
\alias{markov_likelihood}
\title{markov_likelihood}
\usage{
markov_likelihood(sequence, table_kmer_proba, log = TRUE)
}
\arguments{
\item{sequence}{the sequence tested}

\item{table_kmer_proba}{table which contain kmer, count, proba of the model}

\item{log}{}
}
\description{
markov_likelihood
}
