% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_likelihodd_seq.R
\name{markov_likelihood_seq}
\alias{markov_likelihood_seq}
\title{Function to calculate the log likelihood of every sequences}
\usage{
markov_likelihood_seq(seqs, ...)
}
\arguments{
\item{seqs}{any table of multiple colones were the sequences are contain in a colone named "sequence"}

\item{...}{table of probability of a markov model created with function build_markov, model can be named : model1 = modelprobtable.}
}
\value{
a table of columns
  - `...` : all the colomns contain in the original table
  - `sequence` : the DNA sequence (character)
  - `lprob` : the (log)likelihood for each sequence, if the model name is precised, colone name = model name.

a vector of likelihood values
}
\description{
This function calculate the log likelihood of multiples sequences, with a chosen Markov model order.
}
\examples{
Applying with one markov model order :
loglikelihood_table <- markov_likelihood(cpg_sequences, model1 = M1)

Applying with multiple markov model order :
loglikelihood_table <- markov_likelihood_seq(cpg_sequences, model1 = M1) \%>\%
markov_likelihood_seq(model2 = M2)

}
