% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read DNA sequences from a file}
\usage{
read_fasta(file, progress = FALSE, num_threads = readr_threads())
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a
single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip
will be automatically uncompressed. Files starting with  ⁠http://⁠,
⁠https://⁠, ⁠ftp://⁠, or  ⁠ftps://⁠ will be automatically downloaded.
Remote gz files can also be automatically downloaded and decompressed.
Literal data is most useful for examples and tests. To be recognised as
literal data, the input must be either wrapped with I(), be a string
containing at least one new line, or be a vector containing at least one
string with a new line. Using a value of clipboard() will read from the
system clipboard.}

\item{progress}{Display a progress bar? By default it will only display in an
interactive session and not while knitting a document. The automatic
progress bar can be disabled by setting option readr.show_progress to
FALSE.}

\item{num_threads}{The number of processing threads to use for initial
parsing.}
}
\value{
a tibble with two columns
\itemize{
\item \code{id} : the sequence identifier (character)
\item \code{sequence} : the DNA sequence (character)
}
}
\description{
Read DNA sequences from a file
}
\author{
Eric Coissac
}
