% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_table.R
\name{sum_table}
\alias{sum_table}
\title{Calculate the sum of counts for each kmer in multiple kmer tables}
\usage{
sum_table(...)
}
\arguments{
\item{...}{A variable number of beemm_kmer_table objects}
}
\value{
A beemm_kmer_table object with the summed counts for each kmer
}
\description{
Calculate the sum of counts for each kmer in multiple kmer tables
}
\details{
This function takes a variable number of beemm_kmer_table objects
and calculates the sum of counts for each kmer across all tables. It checks
if all the input arguments are beemm_kmer_table objects and if they have the
same kmer size. If not, an error is thrown.
}
\examples{
# Create example kmer tables
table1 <- as_kmer_table(data.frame(kmer = c("AAA", "AAC", "AAG"), count = c(10, 20, 30)))
table2 <- as_kmer_table(data.frame(kmer = c("AAA", "AAT"), count = c(5, 15)))

# Calculate the sum of counts for each kmer
sum_table(table1, table2)

}
