% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\title{HMM Viterbi segmentation of a DNA sequence among two Markov chain models}
\usage{
viterbi(sequence, M0, M1, transitions)
}
\arguments{
\item{sequence}{a character corresponding to the DNA sequence to segment}

\item{M0}{the first Markov model}

\item{M1}{the second Markov model}

\item{transitions}{the transition matrix between models as a 2x2 matrix}
}
\description{
HMM Viterbi segmentation of a DNA sequence among two Markov chain models
}
